requirements.txt
# Python 3.11.13+ required
pytest==7.4.3
requests==2.31.0
python-dotenv==1.0.0
pytest-html==4.1.1
pytest-json-report==1.5.0
pyyaml==6.0.1
aiohttp==3.9.1
pytest-asyncio==0.23.2